requirements.txt

coveralls==1.11.1
google-cloud-secret-manager==0.1.1
pygithub==1.46
